chore: rename python wrapper test. This only houses utility functions used in the PySCF interface
