+++ tests/test_interface_pyscf_tools.py
@@ -5,7 +5,7 @@ import pytest
 
 from fanpy.interface.pyscf_tools import hartreefock, fci_cimatrix, localize
 
-from test_wrapper_python_wrapper import check_data_h2_rhf_sto6g, check_data_lih_rhf_sto6g
+from tests.interface_utils import check_data_h2_rhf_sto6g, check_data_lih_rhf_sto6g
 from utils import find_datafile
